add option to ignore upper limit for some users
This is useful for service users that shouldn't run into any upper
limit. For example, users are paying for using a machine and money
is only deducted from the account when replacement parts have to
be bought, i.e. there can be a lot of money in that account and
user still need to transfer money to the account.